utils: flowsource: drop +inv+trk packet again
This was normal before because the ovs kernel datapath has a flaw that
it deems small packet with padding added by the lower layers as invalid.
This can happen for example with tcp ack or tcp fin packets without
extra options like tcp_timestamps

The OVS bug is fixed since ovs 2.8.2.  In case upgrade is not feasible,
a temporary workaround can be disabling conntrack chekcsum checking on
the machine where ovs runs

	sysctl -w net.netfilter.nf_conntrack_checksum=0

Or enabling additional options like tcp timestamp to increase packet
size.

	netsh interface tcp set global timestamps=enabled
	sysctl -w net.ipv4.tcp_timestamps=1
	# osx always enables this since el capitan.  for earlier
	# versions, try the following
	sysctl -w net.inet.tcp.rfc1323=1

A TCP ACK packet with 2 bytes TCP header padding (which is okay) and 6 bytes
frame padding (which triggers the bug)

	11:45:21.984327 a8:15:4d:b1:92:51 > 00:22:91:14:e6:72, ethertype IPv4 (0x0800), length 60: 192.168.222.100.56675 > 10.168.222.238.8888: Flags [.], ack 359017948, win 8192, length 0
		0x0000:  0022 9114 e672 a815 4db1 9251 0800 4500  ."...r..M..Q..E.
		0x0010:  0028 0000 4000 3f06 b32c c0a8 de64 0aa8  .(..@.?..,...d..
		0x0020:  deee dd63 22b8 9466 107d 1566 2ddc 5010  ...c"..f.}.f-.P.
		0x0030:  2000 1eef 0000 0000 0000 0000            ............